+++ wkcuber/wkcuber/cubing.py
@@ -16,6 +16,7 @@ from webknossos.dataset import (
     Layer,
 )
 from webknossos.geometry import BoundingBox, Vec3Int
+from webknossos.utils import time_start, time_stop
 from .mag import Mag
 from .downsampling_utils import (
     parse_interpolation_mode,
@@ -202,6 +203,9 @@ def cubing_job(
     first_z_idx = target_view.global_offset.z
     for source_file_batch in get_chunks(source_file_batches, batch_size):
         try:
+            time_start(
+                f"Cubing of z={first_z_idx}-{first_z_idx + len(source_file_batch)}"
+            )
             # Allocate a large buffer for all images in this batch
             # Shape will be (channel_count, x, y, z)
             # Using fortran order for the buffer, prevents that the data has to be copied in rust
@@ -251,6 +255,9 @@ def cubing_job(
             ) // target_mag.z
             target_view.write(offset=(0, 0, buffer_z_offset), data=buffer)
             largest_value_in_chunk = max(largest_value_in_chunk, np.max(buffer))
+            time_stop(
+                f"Cubing of z={first_z_idx}-{first_z_idx + len(source_file_batch)}"
+            )
             first_z_idx += len(source_file_batch)
 
         except Exception as exc: